@@ -1,17 +1,16 @@
 ﻿using System;
 
-namespace Authorization.Contracts.Authorization
-{
-    public class UserModel : UserParameters
-    {
-        /// <summary>
-        /// Идентификатор пользователя
-        /// </summary>
-        public Guid Id { get; set; }
+namespace Authorization.Contracts.Authorization;
 
-        /// <summary>
-        /// Дата создания пользователя
-        /// </summary>
-        public DateTime DateOfCreate { get; set; }
-    }
+public class UserModel : UserParameters
+{
+    /// <summary>
+    /// Идентификатор пользователя
+    /// </summary>
+    public Guid Id { get; set; }
+
+    /// <summary>
+    /// Дата создания пользователя
+    /// </summary>
+    public DateTime DateOfCreate { get; set; }
 }
