@@ -8,57 +8,56 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Authorization.Core.Authorization
+namespace Authorization.Core.Authorization;
+
+public class AuthorizationService : IAuthorizationService
 {
-    public class AuthorizationService : IAuthorizationService
+    private readonly IAuthorizationRepository _authorizationRepository;
+    private readonly IMapper _mapper;
+
+    private static JwtOptions _jwtOptions;
+
+    public AuthorizationService(
+        IAuthorizationRepository authorizationRepository,
+        IMapper mapper,
+        IOptions<JwtOptions> jwOptions)
     {
-        private readonly IAuthorizationRepository _authorizationRepository;
-        private readonly IMapper _mapper;
+        _authorizationRepository = authorizationRepository;
+        _mapper = mapper;
+        _jwtOptions = jwOptions.Value;
+    }
 
-        private static JwtOptions _jwtOptions;
+    public async Task<UserModel> GetUserById(Guid userId)
+    {
+        var user = await _authorizationRepository.GetUserById(userId);
 
-        public AuthorizationService(
-            IAuthorizationRepository authorizationRepository,
-            IMapper mapper,
-            IOptions<JwtOptions> jwOptions)
-        {
-            _authorizationRepository = authorizationRepository;
-            _mapper = mapper;
-            _jwtOptions = jwOptions.Value;
-        }
+        return _mapper.Map<UserModel>(user);
+    }
 
-        public async Task<UserModel> GetUserById(Guid userId)
-        {
-            var user = await _authorizationRepository.GetUserById(userId);
+    public async Task<IList<UserModel>> GetUsers()
+    {
+        var users = await _authorizationRepository.GetUsers();
 
-            return _mapper.Map<UserModel>(user);
-        }
+        return _mapper.Map<List<UserModel>>(users);
+    }
 
-        public async Task<IList<UserModel>> GetUsers()
-        {
-            var users = await _authorizationRepository.GetUsers();
+    public async Task<(bool IsSuccess, Guid? UserId)> CreateUser(UserParameters user)
+    {
+        var userEntity = _mapper.Map<UserEntity>(user);
 
-            return _mapper.Map<List<UserModel>>(users);
-        }
+        return await _authorizationRepository.CreateUser(userEntity);
+    }
 
-        public async Task<(bool IsSuccess, Guid? UserId)> CreateUser(UserParameters user)
-        {
-            var userEntity = _mapper.Map<UserEntity>(user);
+    public async Task<AuthenticateResponse> Authorize(AuthenticateParameters authenticateParameters)
+    {
+        var userEntity = await _authorizationRepository
+            .FindUser(authenticateParameters.UserName, authenticateParameters.Password);
 
-            return await _authorizationRepository.CreateUser(userEntity);
-        }
+        if (userEntity == null)
+            return null;
 
-        public async Task<AuthenticateResponse> Authorize(AuthenticateParameters authenticateParameters)
-        {
-            var userEntity = await _authorizationRepository
-                .FindUser(authenticateParameters.UserName, authenticateParameters.Password);
+        var token = new JwtHelper(_jwtOptions).GenerateJwtToken(userEntity);
 
-            if (userEntity == null)
-                return null;
-
-            var token = new JwtHelper(_jwtOptions).GenerateJwtToken(userEntity);
-
-            return new AuthenticateResponse { Token = token };
-        }
+        return new AuthenticateResponse { Token = token };
     }
 }
