@@ -1,14 +1,13 @@
 ﻿using Authorization.Contracts.Authorization;
 using FluentValidation;
 
-namespace Authorization.Validation.Authorization
+namespace Authorization.Validation.Authorization;
+
+public class AuthenticateParametersValidator : AbstractValidator<AuthenticateParameters>
 {
-    public class AuthenticateParametersValidator : AbstractValidator<AuthenticateParameters>
+    /// <param name="userBaseValidator"><see cref="UserBaseValidator"/></param>
+    public AuthenticateParametersValidator(IValidator<UserBase> userBaseValidator)
     {
-        /// <param name="userBaseValidator"><see cref="UserBaseValidator"/></param>
-        public AuthenticateParametersValidator(IValidator<UserBase> userBaseValidator)
-        {
-            Include(userBaseValidator);
-        }
+        Include(userBaseValidator);
     }
 }
