@@ -3,12 +3,11 @@ using System;
 using System.Linq;
 using System.Security.Claims;
 
-namespace Authorization.Controllers
+namespace Authorization.Controllers;
+
+[ApiController]
+[Route("api/[controller]")]
+public class BaseController : ControllerBase
 {
-    [ApiController]
-    [Route("api/[controller]")]
-    public class BaseController : ControllerBase
-    {
-        protected Guid UserId => Guid.Parse(User.Claims.First(c => c.Type == ClaimTypes.NameIdentifier).Value);
-    }
+    protected Guid UserId => Guid.Parse(User.Claims.First(c => c.Type == ClaimTypes.NameIdentifier).Value);
 }
